[#1452] Split CLI commands into separate files
